fix(sdk-ui-dashboard): deduplicate date filter config logic
JIRA: RAIL-4697
