+++ libs/sdk-ui-dashboard/src/presentation/widget/common/configuration/useDateDatasetFilter.ts
@@ -1,0 +1,42 @@
+// (C) 2022 GoodData Corporation
+import { useCallback, useEffect } from "react";
+import { ICatalogDateDataset } from "@gooddata/sdk-model";
+import {
+    selectWidgetDateDatasetAutoSelect,
+    uiActions,
+    useDashboardDispatch,
+    useDashboardSelector,
+} from "../../../../model";
+import { getRecommendedCatalogDateDataset } from "../../../../_staging/dateDatasets/getRecommendedCatalogDateDataset";
+
+export function useDateDatasetFilter(dateDatasets: Readonly<ICatalogDateDataset[]> | undefined) {
+    const dispatch = useDashboardDispatch();
+
+    const isWidgetDateDatasetAutoSelect = useDashboardSelector(selectWidgetDateDatasetAutoSelect);
+
+    useEffect(() => {
+        return () => {
+            // once the config panel disappears set the auto-select flag to false so that editing existing KPIs
+            // does not have it set to true
+            dispatch(uiActions.setWidgetDateDatasetAutoSelect(false));
+        };
+    }, [dispatch]);
+
+    const handleDateDatasetChanged = useCallback(() => {
+        dispatch(uiActions.setWidgetDateDatasetAutoSelect(false));
+    }, [dispatch]);
+
+    /**
+     * Only open the picker if
+     * 1. auto selection happened
+     * 2. there was no recommended dataset
+     *
+     * In that case we want to show the user the picker to pick one of the non-recommended datasets.
+     * Otherwise the preselected recommended dataset is most likely correct so we do not bother the user
+     * with the automatically opened picker.
+     */
+    const shouldOpenDateDatasetPicker =
+        isWidgetDateDatasetAutoSelect && dateDatasets && !getRecommendedCatalogDateDataset(dateDatasets);
+
+    return { handleDateDatasetChanged, shouldOpenDateDatasetPicker };
+}
