+++ libs/sdk-ui-dashboard/src/presentation/widget/insight/configuration/InsightDateDataSetFilter.tsx
@@ -1,5 +1,5 @@
 // (C) 2007-2022 GoodData Corporation
-import React, { useCallback, useEffect } from "react";
+import React, { useEffect } from "react";
 import { DateDatasetFilter } from "../../common";
 import { IInsightWidget, isInsightWidget, widgetRef } from "@gooddata/sdk-model";
 import invariant from "ts-invariant";
@@ -9,13 +9,10 @@ import {
     QueryInsightDateDatasets,
     selectInsightByRef,
     selectIsWidgetLoadingAdditionalDataByWidgetRef,
-    selectWidgetDateDatasetAutoSelect,
-    uiActions,
-    useDashboardDispatch,
     useDashboardQueryProcessing,
     useDashboardSelector,
 } from "../../../../model";
-import { getRecommendedCatalogDateDataset } from "../../../../_staging/dateDatasets/getRecommendedCatalogDateDataset";
+import { useDateDatasetFilter } from "../../common/configuration/useDateDatasetFilter";
 
 export interface IConfigurationPanelProps {
     widget: IInsightWidget;
@@ -34,9 +31,6 @@ export default function InsightDateDataSetFilter({ widget }: IConfigurationPanel
         queryCreator: queryDateDatasetsForInsight,
     });
 
-    const dispatch = useDashboardDispatch();
-
-    const isWidgetDateDatasetAutoSelect = useDashboardSelector(selectWidgetDateDatasetAutoSelect);
     const isLoadingAdditionalData = useDashboardSelector(
         selectIsWidgetLoadingAdditionalDataByWidgetRef(widgetRef(widget)),
     );
@@ -44,36 +38,14 @@ export default function InsightDateDataSetFilter({ widget }: IConfigurationPanel
     const insight = useDashboardSelector(selectInsightByRef(widget.insight));
     invariant(insight, "inconsistent state in InsightDateDataSetFilter");
 
-    const handleDateDatasetChanged = useCallback(() => {
-        dispatch(uiActions.setWidgetDateDatasetAutoSelect(false));
-    }, [dispatch]);
-
     useEffect(() => {
         // use the whole insight to improve cache hits: other calls to this query also use whole insights
         queryDateDatasets(insight);
     }, [queryDateDatasets, insight]);
 
-    useEffect(() => {
-        return () => {
-            // once the config panel disappears set the auto-select flag to false so that editing existing KPIs
-            // does not have it set to true
-            dispatch(uiActions.setWidgetDateDatasetAutoSelect(false));
-        };
-    }, [dispatch]);
-
-    /**
-     * Only open the picker if
-     * 1. auto selection happened
-     * 2. there was no recommended dataset
-     *
-     * In that case we want to show the user the picker to pick one of the non-recommended datasets.
-     * Otherwise the preselected recommended dataset is most likely correct so we do not bother the user
-     * with the automatically opened picker.
-     */
-    const shouldOpenDateDatasetPicker =
-        isWidgetDateDatasetAutoSelect &&
-        result?.dateDatasets &&
-        !getRecommendedCatalogDateDataset(result.dateDatasets);
+    const { handleDateDatasetChanged, shouldOpenDateDatasetPicker } = useDateDatasetFilter(
+        result?.dateDatasets,
+    );
 
     if (isInsightWidget(widget)) {
         return (
