+++ libs/sdk-ui-dashboard/src/presentation/widget/kpi/DefaultKpiConfigurationPanel/KpiWidgetDateDatasetFilter.tsx
@@ -1,5 +1,5 @@
 // (C) 2022 GoodData Corporation
-import React, { useCallback, useEffect } from "react";
+import React, { useEffect } from "react";
 import { IKpiWidget, widgetRef } from "@gooddata/sdk-model";
 
 import { DateDatasetFilter } from "../../common";
@@ -16,7 +16,7 @@ import {
     useDashboardQueryProcessing,
     useDashboardSelector,
 } from "../../../../model";
-import { getRecommendedCatalogDateDataset } from "../../../../_staging/dateDatasets/getRecommendedCatalogDateDataset";
+import { useDateDatasetFilter } from "../../common/configuration/useDateDatasetFilter";
 
 export const KpiWidgetDateDatasetFilter: React.FC<{
     widget: IKpiWidget;
@@ -64,31 +64,9 @@ export const KpiWidgetDateDatasetFilter: React.FC<{
         }
     }, [isWidgetDateDatasetAutoSelect, dispatch, ref, preselectDateDataset]);
 
-    const handleDateDatasetChanged = useCallback(() => {
-        dispatch(uiActions.setWidgetDateDatasetAutoSelect(false));
-    }, [dispatch]);
-
-    /**
-     * Only open the picker if
-     * 1. auto selection happened
-     * 2. there was no recommended dataset
-     *
-     * In that case we want to show the user the picker to pick one of the non-recommended datasets.
-     * Otherwise the preselected recommended dataset is most likely correct so we do not bother the user
-     * with the automatically opened picker.
-     */
-    const shouldOpenDateDatasetPicker =
-        isWidgetDateDatasetAutoSelect &&
-        result?.dateDatasets &&
-        !getRecommendedCatalogDateDataset(result.dateDatasets);
-
-    useEffect(() => {
-        return () => {
-            // once the config panel disappears set the auto-select flag to false so that editing existing KPIs
-            // does not have it set to true
-            dispatch(uiActions.setWidgetDateDatasetAutoSelect(false));
-        };
-    }, [dispatch]);
+    const { handleDateDatasetChanged, shouldOpenDateDatasetPicker } = useDateDatasetFilter(
+        result?.dateDatasets,
+    );
 
     return (
         <div className="gd-kpi-date-dataset-dropdown">
